Removing RestEasy dependency and adding jackson.
As far as I can tell there are now dependencies to RestEasy in the code
base. The dependency to RestEasy was of scope provided which meant that
a project depending on this one would have to define the dependency to
jackson them selves.

Adding jackson as a compile dependency allows project using this one to
not have to add any dependencies.